Refactored gui.onRunCOMRegistrationFixesCommand, to make it fail more gracefully and improve code.
No longer shows success dialog on failure, or when the UAC is canceled or declined.
If there is a Windows error other than cancel of UAC, alert the user and show the error in a dialog.
Slight changes to warning message wording.
Switched from using YES and NO buttons, to YES_NO.
Added a CANCEL button for visual closure and escape key use.
Catch the Windows error that signals user cancel of UAC, and return (as if "NO" had been initially chosen).
Added a docstring.
Added some more logging.