Refactored the COM Registration Fixing Tool to fix bugs and improve UX.
gui.onRunCOMRegistrationFixesCommand:
No longer shows success dialog on failure, or when the UAC is canceled or declined (#12345)..
(Catch the Windows error that signals user cancel of UAC, and return as if "NO" had been initially chosen.)
If there is a Windows error other than cancel of UAC, alert the user and show the error in a dialog.
Rewrote initial dialog message to be more friendly and helpful, and removed warning icon (#12351).
Switched from using YES and NO buttons, to YES_NO.
Added a CANCEL button for visual closure and escape key use.
Added a help button to take users to the tool's user manual section (uses gui.contextHelp.showHelp, but not real context help).
Added a docstring and fully linted code. Added more debug logging.